% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CCI.R
\name{calculate_CCI}
\alias{calculate_CCI}
\title{Calculate CCI index}
\usage{
calculate_CCI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate CCI index
}
\examples{
calculate_CCI()
}
