% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DVI.R
\name{calculate_DVI}
\alias{calculate_DVI}
\title{Calculate DVI index}
\usage{
calculate_DVI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate DVI index
}
\examples{
calculate_DVI(r)
}
