% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GDVI.R
\name{calculate_GDVI}
\alias{calculate_GDVI}
\title{Calculate GDVI index}
\usage{
calculate_GDVI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate GDVI index
}
\examples{
calculate_GDVI(r)
}
