% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GSAVI.R
\name{calculate_GSAVI}
\alias{calculate_GSAVI}
\title{Calculate GSAVI index}
\usage{
calculate_GSAVI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate GSAVI index
}
\examples{
calculate_GSAVI()
}
