% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MSAVI.R
\name{calculate_MSAVI}
\alias{calculate_MSAVI}
\title{Calculate MSAVI index}
\usage{
calculate_MSAVI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate MSAVI index
}
\examples{
calculate_MSAVI()
}
