% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_NDVI.R
\name{calculate_NDVI}
\alias{calculate_NDVI}
\title{Calculate NDVI index}
\usage{
calculate_NDVI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate NDVI index
}
\examples{
calculate_NDVI(r)
}
