% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_new_position.R
\name{calculate_new_position}
\alias{calculate_new_position}
\title{Title}
\usage{
calculate_new_position(start_lon, start_lat, distance, bearing)
}
\arguments{
\item{start_lon}{}

\item{start_lat}{}

\item{distance}{}

\item{bearing}{}
}
\description{
Title
}
