% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_new_tree_position.R
\name{calculate_new_tree_position}
\alias{calculate_new_tree_position}
\title{Calculate New Tree Positions}
\usage{
calculate_new_tree_position(
  plot,
  save,
  crs,
  dataset,
  plot_center_x,
  plot_center_y
)
}
\arguments{
\item{plot}{The plot number for which tree positions are calculated (numeric).}

\item{save}{Logical (\code{TRUE} or \code{FALSE}), indicating whether to save the results to a SHP file.}

\item{crs}{Coordinate reference system in EPSG format (default is 2180, which is the PUWG 1992 system for Poland).}

\item{dataset}{Path to the CSV file containing tree data with columns: \code{plot_id}, \code{azimuth}, \code{distance}.}

\item{plot_center_x}{The x-coordinate of the plot center (numeric).}

\item{plot_center_y}{The y-coordinate of the plot center (numeric).}
}
\value{
A spatial data frame (sf object) with calculated tree positions.
}
\description{
This function calculates the coordinates of trees based on their azimuth and distance from the plot center, and optionally saves the results in \code{sf} format.
}
\examples{
\dontrun{
calculate_new_tree_position(
  plot = 1,
  save = TRUE,
  crs = 2180,
  dataset = "trees_data.csv",
  plot_center_x = 500000,
  plot_center_y = 400000
)
}
}
