% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TGI.R
\name{calculate_TGI}
\alias{calculate_TGI}
\title{Calculate TGI index}
\usage{
calculate_TGI(r)
}
\arguments{
\item{r}{\itemize{
\item stars object
}}
}
\description{
Calculate TGI index
}
\examples{
calculate_TGI()
}
