% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHM_and_TREES_from_XY.R
\name{CHM_and_TREES_from_XY}
\alias{CHM_and_TREES_from_XY}
\title{Creating Canopy Height Models and Tree Detection from XY Coordinates}
\usage{
CHM_and_TREES_from_XY(
  input_X,
  input_Y,
  point_ID,
  d,
  output_folder,
  input_crs_EPSG,
  res_par,
  p2r_par,
  hmin_par,
  ws_par
)
}
\arguments{
\item{input_X}{\itemize{
\item X
}}

\item{input_Y}{\itemize{
\item Y
}}

\item{point_ID}{\itemize{
\item your unique ID
}}

\item{d}{\itemize{
\item distanse for buffer
}}

\item{output_folder}{\itemize{
\item output folder
}}

\item{input_crs_EPSG}{\itemize{
\item EPSG of your CRS data
}}

\item{res_par}{\itemize{
\item resolution of CHM
}}

\item{p2r_par}{\itemize{
\item parametr for p2r() algorithm
}}

\item{hmin_par}{\itemize{
\item minimal height of tree
}}

\item{ws_par}{\itemize{
\item window size
}}
}
\description{
Creating Canopy Height Models and Tree Detection from XY Coordinates
}
\examples{
CHM_and_TREES_from_XY(15.3029620,52.3876487,1,50, output_folder="D:/01_PROJEKTY/01_Powierzchnie_przyrostowe/Uzupelnienie",4326,0.5,0.15,10,5)
}
