% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CHM_from_LAS.R
\name{create_CHM_from_LAS}
\alias{create_CHM_from_LAS}
\title{Title}
\usage{
create_CHM_from_LAS(las_path, res_par, p2r_par, save_chm, output_folder)
}
\arguments{
\item{las_path}{}
}
\description{
Title
}
