% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_AOI.R
\name{define_AOI}
\alias{define_AOI}
\title{Define AOI by setting path to the Shapefile}
\usage{
define_AOI(x)
}
\arguments{
\item{x}{\itemize{
\item Path to the Shapefile
}}
}
\description{
Define AOI by setting path to the Shapefile
}
\examples{
 define_AOI("D:/BDL_Rezerwat_AOI.shp")
}
