% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_coastal_blue.R
\name{define_coastal_blue}
\alias{define_coastal_blue}
\title{Define witch band is Costal Blue}
\usage{
define_coastal_blue(x)
}
\arguments{
\item{x}{}
}
\description{
Define witch band is Costal Blue
}
