% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_download_big_area_rgugik.R
\name{dem_download_big_area_rgugik}
\alias{dem_download_big_area_rgugik}
\title{Title}
\usage{
dem_download_big_area_rgugik(x, y)
}
\arguments{
\item{x}{\itemize{
\item path to your shapefile
}}

\item{y}{\itemize{
\item integer of length 1 or 2, number of grid cells in x and y direction (columns, rows)
}}
}
\description{
Title
}
