% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_tops_in_CHM.R
\name{detect_tops_in_CHM}
\alias{detect_tops_in_CHM}
\title{Title}
\usage{
detect_tops_in_CHM(CHM_path, ws_par, hmin_par, output_folder, save_tops)
}
\arguments{
\item{save_tops}{}
}
\description{
Title
}
