% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foRestools-package.R
\docType{package}
\name{foRestools-package}
\alias{foRestools}
\alias{foRestools-package}
\title{foRestools: Tools for foresters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate vegetation indices and generate statistics and charts for user defined districts
}
\seealso{
Useful links:
\itemize{
  \item \url{https://j-miszczyszyn.github.io/foRestools/}
}

}
\author{
\strong{Maintainer}: First Last \email{jakub.miszczyszyn@urk.edu.pl} (\href{https://orcid.org/0009-0002-8592-946X}{ORCID})

}
\keyword{internal}
