% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersja_temp.R
\name{konwersja_temp}
\alias{konwersja_temp}
\title{Konwersja temperatur}
\usage{
konwersja_temp(temperatura_f)
}
\arguments{
\item{temperatura_f}{}
}
\value{
przeliczona temperatura
}
\description{
Przelicza temperaturę
}
\examples{
konwersja_temp(40)
}
