% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_height_in_clouds.R
\name{normalize_height_in_clouds}
\alias{normalize_height_in_clouds}
\title{Title}
\usage{
normalize_height_in_clouds(las_path, folder, save_las)
}
\arguments{
\item{save_las}{}
}
\description{
Title
}
