% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_cloud.R
\name{thin_cloud}
\alias{thin_cloud}
\title{Title
Thin Point Cloud by Keeping Specific Classes}
\usage{
thin_cloud(las_path, folder, save_las)
}
\arguments{
\item{las_path}{The file path of the LAS file to be thinned.}

\item{folder}{The directory path where the thinned LAS file should be saved, if saving is enabled.}

\item{save_las}{A logical value indicating whether to save the thinned point cloud as a new LAS file.
If TRUE, the thinned point cloud is saved; if FALSE, it is not saved.}
}
\description{
This function reads a LAS file, filters it to retain only specified classes of points (e.g., vegetation),
and optionally saves the thinned point cloud to a new LAS file. The function specifically keeps points
classified as 4 (medium vegetation) and 5 (high vegetation) by default.
}
\details{
This function does not return a value but outputs a LAS file of the thinned point cloud if \code{save_las} is TRUE.
}
\examples{
\dontrun{
thin_cloud("path/to/your/point_cloud.las", "path/to/output/folder", TRUE)
}
}
