% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tops_in_folder_to_shp.R
\name{tops_in_folder_to_shp}
\alias{tops_in_folder_to_shp}
\title{Title}
\usage{
tops_in_folder_to_shp(
  folder_path,
  ws_par,
  hmin_par,
  crs_def,
  output_folder,
  save_tops,
  save_SHP
)
}
\arguments{
\item{save_SHP}{}
}
\description{
Title
}
