% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_new_position.R
\name{calculate_new_position}
\alias{calculate_new_position}
\title{Calculate New Position}
\usage{
calculate_new_position(start_lon, start_lat, distance, bearing)
}
\arguments{
\item{start_lon}{The starting longitude in decimal degrees.}

\item{start_lat}{The starting latitude in decimal degrees.}

\item{distance}{The distance to travel from the starting point, in meters.}

\item{bearing}{The direction of travel in degrees from north (clockwise).}
}
\value{
A list containing the new longitude and latitude in decimal degrees.
}
\description{
This function calculates the new geographical position given a starting longitude, latitude, distance, and bearing.
}
