% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CHM_from_LAS.R
\name{create_CHM_from_LAS}
\alias{create_CHM_from_LAS}
\title{Create Canopy Height Model (CHM) from LAS File}
\usage{
create_CHM_from_LAS(las_path, res_par, p2r_par, save_chm, output_folder)
}
\arguments{
\item{las_path}{The file path of the LAS file from which the CHM is to be created.}

\item{res_par}{The resolution parameter for the CHM raster; specifies the size of each pixel in the output raster.}

\item{p2r_par}{The pixel-to-region parameter used in the p2r algorithm for rasterizing the point cloud.}

\item{save_chm}{A logical value indicating whether to save the generated CHM as a TIFF file.
If TRUE, the CHM is saved; if FALSE, it is not saved.}

\item{output_folder}{The directory path where the CHM TIFF file should be saved if saving is enabled.}
}
\value{
The function returns NULL invisibly. If \code{save_chm} is TRUE, the CHM is saved as a TIFF file.
}
\description{
This function processes a LAS file to generate a Canopy Height Model (CHM) using the specified resolution
and pixel-to-region (p2r) parameters. The resulting CHM can optionally be saved as a TIFF file in the specified output folder.
The function uses the \code{rasterize_canopy} method from the lidR package for rasterizing the LAS point cloud.
}
\examples{
\dontrun{
create_CHM_from_LAS("path/to/your/las_file.las", 0.5, 2, TRUE, "path/to/output/folder")
}
}
