% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_tops_in_CHM.R
\name{detect_tops_in_CHM}
\alias{detect_tops_in_CHM}
\title{Title
Detect Tree Tops in a Canopy Height Model (CHM)}
\usage{
detect_tops_in_CHM(CHM_path, ws_par, hmin_par, output_folder, save_tops)
}
\arguments{
\item{CHM_path}{The file path of the Canopy Height Model (CHM) TIFF file.}

\item{ws_par}{The window size parameter for the LMF algorithm in the lidR package,
indicating the size of the window used to find local maxima.}

\item{hmin_par}{The minimum height parameter for the LMF algorithm in the lidR package,
which specifies the minimum tree height to be considered.}

\item{output_folder}{The directory path where the output shapefile should be saved if saving is enabled.}

\item{save_tops}{A logical value indicating whether to save the detected tree tops as a shapefile.
If TRUE, the tree tops are saved; if FALSE, they are not.}
}
\value{
The function does not return a value but outputs a shapefile of tree tops if \code{save_tops} is TRUE.
}
\description{
This function identifies tree tops in a CHM file using the local maxima filter (LMF) algorithm
provided by the lidR package. The identified tree tops can optionally be saved as a shapefile.
}
\examples{
\dontrun{
detect_tops_in_CHM("path/to/your/CHM.tif", 5, 10, "path/to/output/folder", TRUE)}
}
