% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foRestools-package.R
\docType{package}
\name{foRestools-package}
\alias{foRestools}
\alias{foRestools-package}
\title{foRestools: Tools for foresters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a collection of tools for analyzing forest data and performing remote sensing tasks. Includes functions for creating Canopy Height Models (CHM), detecting trees from XY coordinates, calculating various vegetation indices (CCI, DVI, GDVI, GSAVI, MSAVI, NDVI, TGI), defining Areas of Interest (AOI), and more. Designed to facilitate the processing and analysis of remote sensing data for forest management and research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://j-miszczyszyn.github.io/foRestools/}
}

}
\author{
\strong{Maintainer}: First Last \email{jakub.miszczyszyn@urk.edu.pl} (\href{https://orcid.org/0009-0002-8592-946X}{ORCID})

}
\keyword{internal}
