% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_height_in_clouds.R
\name{normalize_height_in_clouds}
\alias{normalize_height_in_clouds}
\title{Title
Normalize Height of LiDAR Point Cloud}
\usage{
normalize_height_in_clouds(las_path, folder, save_las)
}
\arguments{
\item{las_path}{Path to the input LAS file containing the point cloud data.}

\item{folder}{Path to the folder where the output LAS file will be saved, if saving is enabled.}

\item{save_las}{Logical flag indicating whether to save the normalized point cloud to a new LAS file. Set to TRUE to save.}
}
\value{
las_norm
}
\description{
This function reads a LAS file, normalizes the heights of the point cloud using a TIN surface,
and optionally saves the normalized point cloud to a new LAS file. Only the points classified as
ground (class = 2) are used to compute the TIN surface.
}
\examples{
\dontrun{
# Assuming 'path_to_las' and 'output_folder' are predefined:
normalized_las <- normalize_height_in_clouds("path_to_las", "output_folder", TRUE)
}
}
