% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_download_big_area_rgugik.R
\name{ortho_download_big_area_rgugik}
\alias{ortho_download_big_area_rgugik}
\title{Download Orthophoto for a Large Area from RGUGIK}
\usage{
ortho_download_big_area_rgugik(x, y)
}
\arguments{
\item{x}{The file path to the shapefile defining the area of interest.}

\item{y}{The number of grid cells to divide the area of interest into.}
}
\value{
A dataframe containing the orthophoto data for the specified area.
}
\description{
This function downloads orthophoto data for a large area specified by a shapefile. The area is divided
into a grid of smaller sections to manage the request size. The orthophoto data for each section is requested from the RGUGIK service
and combined into a single dataframe.
}
\examples{
\dontrun{
result <- ortho_download_big_area_rgugik("path/to/your/shapefile.shp", 10)
print(result)}
}
